fixing raycasting() [UNSTABLE]

+++ srcs/raycasting/utils/update_step_and_side_dist_values.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   get_step_and_side_dist_values.c                    :+:      :+:    :+:   */
+/*   update_step_and_side_dist_values.c                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ncoudsi <ncoudsi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/08 12:23:57 by ncoudsi           #+#    #+#             */
-/*   Updated: 2020/09/09 16:10:47 by ncoudsi          ###   ########.fr       */
+/*   Updated: 2020/09/10 11:46:02 by ncoudsi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,22 +17,21 @@ void	update_step_and_side_dist_values()
 	if (ray_dir_x() < 0)
 		{
 			set_step_x(-1);
-			set_side_dist_x((pos_x() - map_pos_x()) * delta_dist_x());
+			set_side_dist_x((pos_x() - (float)map_pos_x()) * delta_dist_x());
 		}
 		else
 		{
 			set_step_x(1);
-			set_side_dist_x((map_pos_x() + 1.0 - pos_x()) * delta_dist_x());
+			set_side_dist_x((float)(map_pos_x() + 1.0f - pos_x()) * delta_dist_x());
 		}
 		if (ray_dir_y() < 0)
 		{
 			set_step_y(-1);
-			set_side_dist_y((pos_y() - map_pos_y()) * delta_dist_y());
+			set_side_dist_y((pos_y() - (float)map_pos_y()) * delta_dist_y());
 		}
 		else
 		{
 			set_step_y(1);
-			set_side_dist_y((map_pos_y() + 1.0 - pos_y()) * delta_dist_y());
+			set_side_dist_y((float)(map_pos_y() + 1.0f - pos_y()) * delta_dist_y());
 		}
 }
-
